Refactor including moving WordSelector and getWordList to their own files.
