+++ src/components/shared/getWordList.ts
@@ -1,0 +1,6 @@
+import * as bip39 from 'bip39';
+
+export default function getWordList(): string[] {
+  const name = bip39.getDefaultWordlist();
+  return bip39.wordlists[name];
+}
